Rename mappings to query
This is just a file rename to make the intent clear.
