+++ src/plugins/profiling/server/routes/query.ts
